fix: move component filtering methods to crates
